- changing datatype from streams::start_time
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@1052 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ sql/hive-0.0.3.sql
@@ -161,7 +161,7 @@ CREATE TABLE IF NOT EXISTS `streams` (
   `codec_name` varchar(255) default NULL,
   `framerate_num` int(11) NOT NULL default '0',
   `framerate_den` int(11) NOT NULL default '0',
-  `start_time` double NOT NULL default '0',
+  `start_time` bigint NOT NULL default '0',
   `duration` bigint NOT NULL default '0',
   `nb_frames` bigint NOT NULL default '0',
   `time_base_num` int(11) NOT NULL default '0',
